Moved EffectControllersMetadata.cs to better represent it's role
