+++ Source/Waterfall/EffectControllers/WaterfallController.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using UnityEngine;
-using Waterfall.UI.EffectControllersUI;
+using Waterfall.EffectControllers;
 
 namespace Waterfall
 {
