+++ Source/Waterfall/Modules/ModuleWaterfallFX.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
-using Waterfall.UI.EffectControllersUI;
+using Waterfall.EffectControllers;
 
 namespace Waterfall
 {
