+++ Source/Waterfall/Waterfall.csproj
@@ -90,6 +90,7 @@
     <Compile Include="AssetLibrary\WaterfallAsset.cs" />
     <Compile Include="EffectControllers\AtmosphereDensityController.cs" />
     <Compile Include="EffectControllers\CustomController.cs" />
+    <Compile Include="EffectControllers\EffectControllersMetadata.cs" />
     <Compile Include="EffectControllers\WaterfallController.cs" />
     <Compile Include="EffectControllers\EngineEventController.cs" />
     <Compile Include="EffectControllers\GimbalController.cs" />
@@ -128,7 +129,6 @@
     <Compile Include="Modules\ModuleWaterfallSmoke.cs" />
     <Compile Include="Serializer\SerializedData.cs" />
     <Compile Include="Settings.cs" />
-    <Compile Include="UI\EffectControllersUI\EffectControllersMetadata.cs" />
     <Compile Include="UI\EffectControllersUI\IEffectControllerUIOptions.cs" />
     <Compile Include="UI\EffectControllersUI\AtmosphereDensityControllerUIOptions.cs" />
     <Compile Include="UI\EffectControllersUI\CustomControllerUIOptions.cs" />
